Reorganize Yarel "softwre"
The single entry-point *yarel-language-structure* now exists.
It is generated as an Eclipse java configuration from the instance of Eclipse that defines Yarel as a DSL.

The entry point yarel-language-structure contains the java project language-structure.
It is generated as an Eclipse java project from the instance of Eclipse that works as IDE for Yarel.

The project language-structure contains:

- **yarel**. It is a source folder of yarel files with extension 'rl'. It contains hierachies of folders that must be thought of as hierarchies of packages.

For example, if a file has name *a_b_c_name.rl*, then it is assumed the java hierarchy of packages *a.b.c* contains it.

This convention will become obsolete as soon as we shall implement a hierarchical implementation of Yarel modules, i.e. of Yarel packages.

- **src**. It contains the sources of java classes that Yarel sources are compiled to. For example, the file a_b_c_name.rl becomes a java package a_b_c_name which contains a class for every function that a_b_c_name.rl declares and defines, which their inverse included, of course.

- **yarel_from_java**. It is meant to access the classes in src, by further java classes, thay may take advantaqge of classes, and theier inverses, that we produce by programming in Yarel and by compiling them in java. Typically, it contains packages with JUnit classes that test classes we find in src. Every class inthe packages of yarel_from_java can be generated starting from a very rough testing kernel that the compiler generates for every class in the packages of src.

For example, let us assume that the package a_b_c_name exists in src. Then a_b_c_name contains the class a_b_c_namePlayWith.java that makes available calls to the classes that the package a_b_c_name contains. We can generate a JUnit class, relative to a_b_c_name in yarel_from_java and use a_b_c_namePlayWith.java as an initial pattern for interesting testing of the classes, i.e. Yarel functions, of a_b_c_name.

This is another part of Yarel implementation whose usability we plan to improve.